VCF roundtrip tests
Test icechunk transaction
